pyproject.toml
[tool.poetry]
name = "nonebot-plugin-wx4"
version = "1.2.1"
description = ""
authors = ["pasumao"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
nonebot2 = "^2.0.0"
requests = "^2.0.0"
nonebot-adapter-onebot = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
